% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{cgem_blue}
\alias{cgem_blue}
\title{HEX Code of the CGEM Blue}
\usage{
cgem_blue()
}
\value{
The HEX code of the CGEM blue.
}
\description{
Access the HEX code of the blue used by CGEM
}
\examples{

barplot(
  1:3,
  col = cgem_blue()
)

}
\seealso{
Other CGEM Colours: 
\code{\link{cgem_red}()}
}
\concept{CGEM Colours}
