% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{cgem_colours}
\alias{cgem_colours}
\title{HEX Codes of the CGEM Colours}
\usage{
cgem_colours()
}
\value{
The HEX codes of the CGEM colours.
}
\description{
Access the HEX codes of the colours used in the CGEM logo.
}
\examples{

barplot(
  1:2,
  col = cgem_colours()
)

}
\seealso{
Other Palettes: 
\code{\link{cgem_colours_n}()},
\code{\link{cruk_colours_n}()},
\code{\link{cruk_colours}()},
\code{\link{ebio_colours_n}()},
\code{\link{ebio_colours}()},
\code{\link{ukri_blues_n}()},
\code{\link{ukri_blues}()},
\code{\link{ukri_purples_n}()},
\code{\link{ukri_purples}()}
}
\concept{Palettes}
