% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{cgem_colours_n}
\alias{cgem_colours_n}
\title{Interpolated HEX Codes of the CGEM Colours}
\usage{
cgem_colours_n(
  n,
  colours = cgem_colours(),
  bias = 1,
  space = "Lab",
  interpolate = "linear",
  alpha = FALSE,
  ...
)
}
\arguments{
\item{n}{The number of colours to return.}

\item{colours}{The colours to interpolate between. By default these are set
to the colours from the CGEM logo.}

\item{bias}{A positive number. Higher values will give more widely spaced
colours at the higher end of the colour spectrum.}

\item{space}{A character string indicating whether the colours should be
interpolated in RGB or CIE Lab colour spaces.}

\item{interpolate}{A character indicating whether a spline or linear
interpolation method should be used.}

\item{alpha}{A logical value indicating whether an alpha channel should be
returned.}

\item{...}{Arguments to be passed to \code{grDevices::colorRampPalette}.}
}
\value{
A character vector of colours iterpolated from the colours in the CGEM
logo.
}
\description{
Access HEX codes of colours that have been interpolated between those used in
the CGEM logo.
}
\examples{

barplot(
  1:5,
  col = cgem_colours_n(5)
)

}
\seealso{
Other Palettes: 
\code{\link{cgem_colours}()},
\code{\link{cruk_colours_n}()},
\code{\link{cruk_colours}()},
\code{\link{ukri_blues_n}()},
\code{\link{ukri_blues}()},
\code{\link{ukri_purples_n}()},
\code{\link{ukri_purples}()}
}
\concept{Palettes}
