% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{cgem_red}
\alias{cgem_red}
\title{HEX Code of the CGEM Red}
\usage{
cgem_red()
}
\value{
The HEX code of the CGEM red.
}
\description{
Access the HEX code of the red used by CGEM
}
\examples{

barplot(
  1:3,
  col = cgem_red()
)

}
\seealso{
Other CGEM Colours: 
\code{\link{cgem_blue}()}
}
\concept{CGEM Colours}
