% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{cruk_colours}
\alias{cruk_colours}
\title{HEX Codes of the CRUK Colours}
\usage{
cruk_colours()
}
\value{
The HEX codes of the CRUK colours.
}
\description{
Access the HEX codes of the colours used in the CRUK logo.
}
\examples{

barplot(
  1:3,
  col = cruk_colours()
)

}
\seealso{
Other Palettes: 
\code{\link{cgem_colours_n}()},
\code{\link{cgem_colours}()},
\code{\link{cruk_colours_n}()},
\code{\link{ebio_colours_n}()},
\code{\link{ebio_colours}()},
\code{\link{ukri_blues_n}()},
\code{\link{ukri_blues}()},
\code{\link{ukri_purples_n}()},
\code{\link{ukri_purples}()}
}
\concept{Palettes}
