% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{cruk_colours_palette_func_gen}
\alias{cruk_colours_palette_func_gen}
\title{Palette Generating Function For CRUK Colours}
\usage{
cruk_colours_palette_func_gen(
  colours = cruk_colours(),
  bias = 1,
  space = "Lab",
  interpolate = "linear",
  alpha = FALSE,
  ...
)
}
\arguments{
\item{colours}{The colours to interpolate between. By default these are set
to the blues from the UKRI logo.}

\item{bias}{A positive number. Higher values will give more widely spaced
colours at the higher end of the colour spectrum.}

\item{space}{A character string indicating whether the colours should be
interpolated in RGB or CIE Lab colour spaces.}

\item{interpolate}{A character indicating whether a spline or linear
interpolation method should be used.}

\item{alpha}{A logical value indicating whether an alpha channel should be
returned.}

\item{...}{Arguments to be passed to \code{grDevices::colorRampPalette}.}
}
\value{
A function to generate palettes that interpolate the CRUK colours.
}
\description{
Palette Generating Function For CRUK Colours
}
\seealso{
Other Helper Functions: 
\code{\link{cgem_colours_palette_func_gen}()},
\code{\link{ebio_colours_palette_func_gen}()},
\code{\link{ukri_blues_palette_func_gen}()},
\code{\link{ukri_purples_palette_func_gen}()}
}
\concept{Helper Functions}
