% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{cruk_dark_blue}
\alias{cruk_dark_blue}
\title{HEX Code of the CRUK Dark Blue}
\usage{
cruk_dark_blue()
}
\value{
The HEX code of the CRUK dark blue.
}
\description{
Access the HEX code of the dark blue used by CRUK.
}
\examples{

barplot(
  1:3,
  col = cruk_dark_blue()
)

}
\seealso{
Other CRUK Colours: 
\code{\link{cruk_grey}()},
\code{\link{cruk_light_blue}()},
\code{\link{cruk_pink}()}
}
\concept{CRUK Colours}
