% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{cruk_pink}
\alias{cruk_pink}
\title{HEX Code of the CRUK Pink}
\usage{
cruk_pink()
}
\value{
The HEX code of the CRUK pink.
}
\description{
Access the HEX code of the pink used by CRUK.
}
\examples{

barplot(
  1:3,
  col = cruk_pink()
)

}
\seealso{
Other CRUK Colours: 
\code{\link{cruk_dark_blue}()},
\code{\link{cruk_grey}()},
\code{\link{cruk_light_blue}()}
}
\concept{CRUK Colours}
