% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ebio_blue}
\alias{ebio_blue}
\title{HEX Code of the EastBIO Blue}
\usage{
ebio_blue()
}
\value{
The HEX code of the EastBIO blue.
}
\description{
Access the HEX code of the blue used by EastBIO
}
\examples{

barplot(
  1:3,
  col = ebio_blue()
)

}
\seealso{
Other EBIO Colours: 
\code{\link{ebio_green}()},
\code{\link{ebio_orange}()},
\code{\link{ebio_red}()}
}
\concept{EBIO Colours}
