% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{ebio_colours}
\alias{ebio_colours}
\title{HEX Codes of the EastBIO Colours}
\usage{
ebio_colours()
}
\value{
The HEX codes of the EastBIO colours.
}
\description{
Access the HEX codes of the colours used in the EastBIO logo.
}
\examples{

barplot(
  1:4,
  col = ebio_colours()
)


}
\seealso{
Other Palettes: 
\code{\link{cgem_colours_n}()},
\code{\link{cgem_colours}()},
\code{\link{cruk_colours_n}()},
\code{\link{cruk_colours}()},
\code{\link{ebio_colours_n}()},
\code{\link{ukri_blues_n}()},
\code{\link{ukri_blues}()},
\code{\link{ukri_purples_n}()},
\code{\link{ukri_purples}()}
}
\concept{Palettes}
