% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ebio_green}
\alias{ebio_green}
\title{HEX Code of the EastBIO Green}
\usage{
ebio_green()
}
\value{
The HEX code of the EastBIO green.
}
\description{
Access the HEX code of the green used by EastBIO
}
\examples{

barplot(
  1:3,
  col = ebio_green()
)

}
\seealso{
Other EBIO Colours: 
\code{\link{ebio_blue}()},
\code{\link{ebio_orange}()},
\code{\link{ebio_red}()}
}
\concept{EBIO Colours}
