% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ebio_orange}
\alias{ebio_orange}
\title{HEX Code of the EastBIO Orange}
\usage{
ebio_orange()
}
\value{
The HEX code of the EastBIO orange.
}
\description{
Access the HEX code of the orange used by EastBIO
}
\examples{

barplot(
  1:3,
  col = ebio_orange()
)

}
\seealso{
Other EBIO Colours: 
\code{\link{ebio_blue}()},
\code{\link{ebio_green}()},
\code{\link{ebio_red}()}
}
\concept{EBIO Colours}
