% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ebio_red}
\alias{ebio_red}
\title{HEX Code of the EastBIO red}
\usage{
ebio_red()
}
\value{
The HEX code of the EastBIO red.
}
\description{
Access the HEX code of the red used by EastBIO
}
\examples{

barplot(
  1:3,
  col = ebio_red()
)

}
\seealso{
Other EBIO Colours: 
\code{\link{ebio_blue}()},
\code{\link{ebio_green}()},
\code{\link{ebio_orange}()}
}
\concept{EBIO Colours}
