% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_palette.R
\name{hgu_dark}
\alias{hgu_dark}
\title{Hex Value Of HGU Dark Blue}
\usage{
hgu_dark()
}
\value{
A character string of the hex value of the HGU dark blue.
}
\description{
Hex Value Of HGU Dark Blue
}
\examples{
colour <- hgu_dark()
colour
}
