% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{hgu_light}
\alias{hgu_light}
\title{Hex Value Of HGU Light Blue}
\usage{
hgu_light()
}
\value{
A character string of the hex value of the HGU light blue.
}
\description{
Hex Value Of HGU Light Blue
}
\examples{
colour <- hgu_light()
colour
}
