% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{hgu_mid}
\alias{hgu_mid}
\title{Hex Value Of HGU Mid Blue}
\usage{
hgu_mid()
}
\value{
A character string of the hex value of the HGU mid blue.
}
\description{
Hex Value Of HGU Mid Blue
}
\examples{
colour <- hgu_mid()
colour
}
