% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{hgu_palette}
\alias{hgu_palette}
\title{Hex Value Colour Palette}
\usage{
hgu_palette()
}
\value{
A character vector of colours in the HGU logo.
}
\description{
Hex Value Colour Palette
}
\examples{
colour_palette <- hgu_palette
colour_palette
}
