% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{hgu_palette_func_gen}
\alias{hgu_palette_func_gen}
\title{Generate The Palette Function For The HGU Blues}
\usage{
hgu_palette_func_gen(
  colours = hgu_palette(),
  bias = 1,
  space = "Lab",
  interpolate = "linear",
  alpha = FALSE,
  ...
)
}
\arguments{
\item{colours}{Colours to interpolate; by default set as the HGU blues colour palette.}

\item{bias}{A positive number; Higher values give more widely spaced colours at the high end.}

\item{space}{A character string; Interpolation in RGB or CIE Lab colour space.}

\item{interpolate}{Use a spline or linear interpolation.}

\item{alpha}{Logical; Should alpha channel (opacity) values be returned?}

\item{...}{}
}
\value{
A function to create discrete colour palettes
}
\description{
Generate The Palette Function For The HGU Blues
}
\examples{
palette_func <- hgu_palette_func()
palette_func(5)
}
