% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{hgu_palette_n}
\alias{hgu_palette_n}
\title{Get Any Number Of Colours From The MRC HGU Blue Palette}
\usage{
hgu_palette_n(n)
}
\arguments{
\item{n}{The number of colours to return}
}
\value{
A character vecctor of colours
}
\description{
Get Any Number Of Colours From The MRC HGU Blue Palette
}
\examples{
hgu_palette_n(10)
}
