% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_dark_blue}
\alias{mrc_dark_blue}
\title{Hex Value Of MRC Dark Blue}
\usage{
mrc_dark_blue()
}
\value{
A character string of the hex value of the MRC dark blue.
}
\description{
Hex Value Of MRC Dark Blue
}
\examples{
colour <- mrc_dark_blue()
colour
}
