% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_darkish_blue}
\alias{mrc_darkish_blue}
\title{Hex Value Of MRC Darkish Blue}
\usage{
mrc_darkish_blue()
}
\value{
A character string of the hex value of the MRC darkish blue.
}
\description{
Hex Value Of MRC Darkish Blue
}
\examples{
colour <- mrc_darkish_blue()
colour
}
