% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_gray}
\alias{mrc_gray}
\title{Hex Value Of MRC Gray}
\usage{
mrc_gray()
}
\value{
A character string of the hex value of the MRC gray.
}
\description{
Hex Value Of MRC Gray
}
\examples{
colour <- mrc_gray()
colour
}
