% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_green}
\alias{mrc_green}
\title{Hex Value Of MRC Green}
\usage{
mrc_green()
}
\value{
A character string of the hex value of the MRC green.
}
\description{
Hex Value Of MRC Green
}
\examples{
colour <- mrc_green()
colour
}
