% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_grey}
\alias{mrc_grey}
\title{Hex Value Of MRC Grey}
\usage{
mrc_grey()
}
\value{
A character string of the hex value of the MRC grey.
}
\description{
Hex Value Of MRC Grey
}
\examples{
colour <- mrc_grey()
colour
}
