% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_mid_blue}
\alias{mrc_mid_blue}
\title{Hex Value Of MRC Mid Blue}
\usage{
mrc_mid_blue()
}
\value{
A character string of the hex value of the MRC mid blue.
}
\description{
Hex Value Of MRC Mid Blue
}
\examples{
colour <- mrc_mid_blue()
colour
}
