% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_red}
\alias{mrc_red}
\title{Hex Value Of MRC Red}
\usage{
mrc_red()
}
\value{
A character string of the hex value of the MRC red.
}
\description{
Hex Value Of MRC Red
}
\examples{
colour <- mrc_red()
colour
}
