% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{mrc_yellow}
\alias{mrc_yellow}
\title{Hex Value Of MRC Yellow}
\usage{
mrc_yellow()
}
\value{
A character string of the hex value of the MRC yellow.
}
\description{
Hex Value Of MRC Yellow
}
\examples{
colour <- mrc_yellow()
colour
}
