% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_colour_continuous.R
\name{scale_colour_cruk_c}
\alias{scale_colour_cruk_c}
\title{CRUK Logo Colour Scales (Continuous)}
\usage{
scale_colour_cruk_c(
  ...,
  low = cruk_dark_blue(),
  mid = cruk_light_blue(),
  high = cruk_pink(),
  midpoint = NULL,
  space = "Lab",
  na.value = cruk_grey(),
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_colour_gradient} or
\code{ggplot2::scale_colour_gradient2}.}

\item{low}{Colour for the low end of the gradient.}

\item{mid}{Colour for the midpoit of the gradient.}

\item{high}{Colour for the high end of the gradient.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Diverging
scale only used if the midpoint is not equal to \code{NULL}, which is the default.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
A layer that can be added to \code{ggplot2} plots to change the colour aesthetic
to interpolate values from the blues used in the CRUK logo.
}
\examples{

library(ggplot2)

ggplot(
  data.frame(
    x = runif(1500),
    y = runif(1500)
  )
) +
  aes(x = x, y = y, colour = x) +
  geom_point() + theme_classic() +
  guides(colour = "none") +
  coord_equal() +
  scale_colour_cruk_c()

ggplot(
  data.frame(
    x = runif(1500),
    y = runif(1500)
  )
) +
  aes(x = x, y = y, colour = x) +
  geom_point() + theme_classic() +
  guides(colour = "none") +
  coord_equal() +
  scale_colour_cruk_c(midpoint = 0.5)

}
