% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_colour_gradient.R
\name{scale_colour_gradient2_hgu}
\alias{scale_colour_gradient2_hgu}
\title{Scale Colours With A Gradient}
\usage{
scale_colour_gradient2_hgu(
  ...,
  low = hgu_dark(),
  mid = hgu_mid(),
  high = hgu_light(),
  midpoint = 0,
  space = "Lab",
  na.value = mrc_grey(),
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{low}{Colour for the low end of the gradient.}

\item{mid}{Colour for the midpoit of the gradient.}

\item{high}{Colour for the high end of the gradient.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Defaults to 0.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
Scale Colours With A Gradient
}
\examples{
ggplot(
  data = data.frame(x = runif(100), y = runif(100)),
  mapping = aes(x = x, y = y, colour = x)
) +
  geom_point() + theme_classic() +
  scale_colour_gradient2_hgu(midpoint = 0.5)
}
