% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_colours.R
\name{scale_colour_gradient_hgu}
\alias{scale_colour_gradient_hgu}
\title{Scale Colours With A Gradient}
\usage{
scale_colour_gradient_hgu(
  ...,
  low = "#2D2E5F",
  high = "#54BBD2",
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{aesthetics}{}
}
\description{
Scale Colours With A Gradient
}
