% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_colour_discrete.R
\name{scale_colour_hgu2}
\alias{scale_colour_hgu2}
\title{Assign Discrete HGU 2 Colour Palette}
\usage{
scale_colour_hgu2(
  ...,
  values = c(hgu_dark(), hgu_light()),
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = mrc_grey()
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{values}{A character vector of colour values to map data to.}

\item{aesthetics}{The aesthetic to apply this scale to.}

\item{breaks}{Values for beaks (see \code{ggplot2::scale_colour_manual} for details).}

\item{na.value}{Colour to be used for missing values.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
Assign Discrete HGU 2 Colour Palette
}
\examples{
library(ggplot2)
ggplot(
  data = data.frame(x = runif(60), y = runif(60), colour = rep(c("A", "B"), 30)),
  mapping = aes(x = x, y = y, colour = colour)
) +
  geom_point() + theme_classic() +
  scale_colour_hgu2()
}
