% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_discrete.R
\name{scale_colour_hgu_3col}
\alias{scale_colour_hgu_3col}
\title{Assign Discrete HGU 3 Colour Palette}
\usage{
scale_colour_hgu_3col(
  ...,
  values = c(hgu_dark(), hgu_mid(), hgu_light()),
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50"
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{values}{A character vector of colour values to map data to.}

\item{aesthetics}{The aesthetic to apply this scale to.}

\item{breaks}{Values for beaks (see \code{ggplot2::scale_colour_manual} for details).}

\item{na.value}{Colour to be used for missing values.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
Assign Discrete HGU 3 Colour Palette
}
\examples{
library(tidyverse)
tibble(
  x = rnorm(60),
  y = rnorm(60),
  col = rep(c("a", "b", "c"), 20)
) \%>\%
  ggplot(aes(x = x, y = y, colour = col)) +
  geom_point() +
  scale_colour_hgu_3col() +
  theme_classic()
}
