% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_colour_continuous.R
\name{scale_colour_ukri_c}
\alias{scale_colour_ukri_c}
\title{UKRI Blues Colour Scales (Continuous)}
\usage{
scale_colour_ukri_c(
  ...,
  colours = ukri_blues(),
  midpoint = NULL,
  space = "Lab",
  na.value = cruk_grey(),
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_colour_gradient} or
\code{ggplot2::scale_colour_gradient2}.}

\item{colours}{The colour palette to use for the colours of the plot. By
default this is the blues from the UKRI logo. Setting this to \code{"purple"}
will switch this to the UKRI purples.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Diverging
scale only used if the midpoint is not equal to \code{NULL}, which is the default.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
A layer that can be added to \code{ggplot2} plots to change the colour aesthetic
to interpolate values from the blues used in the UKRI logo.
}
\examples{

library(ggplot2)

ggplot(
  data.frame(
    x = runif(1500),
    y = runif(1500)
  )
) +
  aes(x = x, y = y, colour = x) +
  geom_point() + theme_classic() +
  guides(colour = "none") +
  coord_equal() +
  scale_colour_ukri_c()

ggplot(
  data.frame(
    x = runif(1500),
    y = runif(1500)
  )
) +
  aes(x = x, y = y, colour = x) +
  geom_point() + theme_classic() +
  guides(colour = "none") +
  coord_equal() +
  scale_colour_ukri_c(
    colours = "purple"
  )

}
\seealso{
Other Layer: 
\code{\link{scale_colour_cgem_c}()},
\code{\link{scale_colour_cgem_d}()},
\code{\link{scale_colour_cruk_c}()},
\code{\link{scale_colour_cruk_d}()},
\code{\link{scale_colour_ukri_d}()},
\code{\link{scale_fill_cgem_c}()},
\code{\link{scale_fill_cgem_d}()},
\code{\link{scale_fill_cruk_c}()},
\code{\link{scale_fill_cruk_d}()},
\code{\link{scale_fill_ukri_c}()},
\code{\link{scale_fill_ukri_d}()}
}
\concept{Layer}
