% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_fill_continuous.R
\name{scale_fill_cruk_c}
\alias{scale_fill_cruk_c}
\title{CRUK Logo Fill Scales (Continuous)}
\usage{
scale_fill_cruk_c(
  ...,
  colours = cruk_colours(),
  midpoint = NULL,
  space = "Lab",
  na.value = cruk_grey(),
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_colour_gradient} or
\code{ggplot2::scale_colour_gradient2}.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Diverging
scale only used if the midpoint is not equal to \code{NULL}, which is the default.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}

\item{low}{Colour for the low end of the gradient.}

\item{mid}{Colour for the midpoint of the gradient.}

\item{high}{Colour for the high end of the gradient.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
CRUK Logo Fill Scales (Continuous)
}
\examples{

library(ggplot2)

ggplot(
  data.frame(
    x = rnorm(10000),
    y = rnorm(10000)
  )
) +
  aes(x = x, y = y) +
  geom_hex() + theme_classic() +
  guides(fill = "none") + coord_equal() +
  scale_fill_cruk_c()

}
\seealso{
Other Layer: 
\code{\link{scale_colour_cruk_c}()},
\code{\link{scale_colour_cruk_d}()},
\code{\link{scale_colour_ukri_c}()},
\code{\link{scale_colour_ukri_d}()},
\code{\link{scale_fill_cruk_d}()},
\code{\link{scale_fill_ukri_c}()},
\code{\link{scale_fill_ukri_d}()}
}
\concept{Layer}
