% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_fill_gradient.R
\name{scale_fill_gradient2_hgu}
\alias{scale_fill_gradient2_hgu}
\title{Scale Fill With A Gradient}
\usage{
scale_fill_gradient2_hgu(
  ...,
  low = hgu_dark(),
  mid = hgu_mid(),
  high = hgu_light(),
  midpoint = 0,
  space = "Lab",
  na.value = mrc_grey(),
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{low}{Colour for the low end of the gradient.}

\item{mid}{Colour for the midpoit of the gradient.}

\item{high}{Colour for the high end of the gradient.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Defaults to 0.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
Scale Fill With A Gradient
}
\examples{
library(ggplot2)
ggplot(
  data = data.frame(x = rnorm(10000), y = rnorm(10000)),
  mapping = aes(x = x, y = y)
) +
  geom_hex(bins = 40) + theme_classic() +
  scale_fill_gradient2_hgu(midpoint = 25)
}
