% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgu_fill_discrete.R
\name{scale_fill_hgu2}
\alias{scale_fill_hgu2}
\title{Assign Discrete HGU 2 Colour Fill Palette}
\usage{
scale_fill_hgu2(
  ...,
  values = c(hgu_dark(), hgu_light()),
  aesthetics = "fill",
  breaks = ggplot2::waiver(),
  na.value = mrc_grey()
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{values}{A character vector of colour values to map data to.}

\item{aesthetics}{The aesthetic to apply this scale to.}

\item{breaks}{Values for beaks (see \code{ggplot2::scale_colour_manual} for details).}

\item{na.value}{Colour to be used for missing values.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
Assign Discrete HGU 2 Colour Fill Palette
}
\examples{
library(tidyverse)
tibble(
  x = rnorm(30),
  y = runif(30)
) \%>\%
  ggplot(aes(x=x,y=y)) +
  geom_density2d_filled(bins = 2) +
  theme_classic() +
  scale_fill_hgu2()
}
