% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{ukri_blues}
\alias{ukri_blues}
\title{HEX Codes of the UKRI Blues}
\usage{
ukri_blues()
}
\value{
The HEX codes of the UKRI blues.
}
\description{
Access the HEX codes of the blues used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_blues()
)

}
\seealso{
Other Palette: 
\code{\link{cruk_colours_n}()},
\code{\link{cruk_colours}()},
\code{\link{ukri_blues_n}()}
}
\concept{Palette}
