% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ukri_dark_blue}
\alias{ukri_dark_blue}
\title{HEX Code of the UKRI Dark Blue}
\usage{
ukri_dark_blue()
}
\value{
The HEX code of the UKRI dark blue.
}
\description{
Access the HEX code of the dark blue used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_dark_blue()
)

}
\seealso{
Other UKRI Colours: 
\code{\link{ukri_light_blue}()},
\code{\link{ukri_mid_blue}()},
\code{\link{ukri_orange}()}
}
\concept{UKRI Colours}
