% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ukri_light_blue}
\alias{ukri_light_blue}
\title{HEX Code of the UKRI Light Blue}
\usage{
ukri_light_blue()
}
\value{
The HEX code of the UKRI light blue.
}
\description{
Access the HEX code of the light blue used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_light_blue()
)

}
\seealso{
Other Colour: 
\code{\link{cruk_dark_blue}()},
\code{\link{cruk_grey}()},
\code{\link{cruk_light_blue}()},
\code{\link{cruk_pink}()},
\code{\link{ukri_dark_blue}()},
\code{\link{ukri_mid_blue}()},
\code{\link{ukri_orange}()}
}
\concept{Colour}
