% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ukri_light_purple}
\alias{ukri_light_purple}
\title{HEX Code of the UKRI Light Purple}
\usage{
ukri_light_purple()
}
\value{
The HEX code of the UKRI light purple.
}
\description{
Access the HEX code of the light purple used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_light_purple()
)

}
\seealso{
Other UKRI Colours: 
\code{\link{ukri_dark_blue}()},
\code{\link{ukri_dark_purple}()},
\code{\link{ukri_light_blue}()},
\code{\link{ukri_mid_blue}()},
\code{\link{ukri_mid_purple}()},
\code{\link{ukri_orange}()}
}
\concept{UKRI Colours}
