% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{ukri_purples}
\alias{ukri_purples}
\title{HEX Codes of the UKRI Purples}
\usage{
ukri_purples()
}
\value{
The HEX codes of the UKRI purples.
}
\description{
Access the HEX codes of the purples used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_purples()
)

}
\seealso{
Other Palettes: 
\code{\link{cgem_colours_n}()},
\code{\link{cgem_colours}()},
\code{\link{cruk_colours_n}()},
\code{\link{cruk_colours}()},
\code{\link{ebio_colours_n}()},
\code{\link{ebio_colours}()},
\code{\link{ukri_blues_n}()},
\code{\link{ukri_blues}()},
\code{\link{ukri_purples_n}()}
}
\concept{Palettes}
