% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{cruk_grey}
\alias{cruk_grey}
\title{HEX Code of the CRUK Grey}
\usage{
cruk_grey()
}
\value{
The HEX code of the CRUK grey.
}
\description{
Access the HEX code of the grey used by CRUK.
}
\examples{

barplot(
  1:3,
  col = cruk_grey()
)

}
