% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_fill_continuous.R
\name{scale_fill_cruk_c}
\alias{scale_fill_cruk_c}
\title{CRUK Logo Fill Scales (Continuous)}
\usage{
scale_fill_cruk_c(
  ...,
  low = cruk_dark_blue(),
  mid = cruk_light_blue(),
  high = cruk_pink(),
  midpoint = NULL,
  space = "Lab",
  na.value = cruk_grey(),
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_colour_gradient} or
\code{ggplot2::scale_colour_gradient2}.}

\item{low}{Colour for the low end of the gradient.}

\item{mid}{Colour for the midpoint of the gradient.}

\item{high}{Colour for the high end of the gradient.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Diverging
scale only used if the midpoint is not equal to \code{NULL}, which is the default.}

\item{space}{Colour space in which to calculate the gradient.}

\item{na.value}{Colour to be used for missing values.}

\item{guide}{Type of legend.}

\item{aesthetics}{The aesthetic to apply this scale to.}
}
\value{
A layer that can be added to a ggplot2 object.
}
\description{
CRUK Logo Fill Scales (Continuous)
}
