% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ukri_mid_blue}
\alias{ukri_mid_blue}
\title{HEX Code of the UKRI Mid Blue}
\usage{
ukri_mid_blue()
}
\value{
The HEX code of the UKRI mid blue.
}
\description{
Access the HEX code of the mid blue used in the UKRI logo.
}
\examples{

barplot(
  1:3,
  col = ukri_mid_blue()
)

}
\seealso{
Other UKRI Colours: 
\code{\link{ukri_dark_blue}()},
\code{\link{ukri_dark_purple}()},
\code{\link{ukri_light_blue}()},
\code{\link{ukri_light_purple}()},
\code{\link{ukri_mid_purple}()},
\code{\link{ukri_orange}()}
}
\concept{UKRI Colours}
