% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_codes.R
\name{ukri_orange}
\alias{ukri_orange}
\title{HEX Code of the UKRI Light Blue}
\usage{
ukri_orange()
}
\value{
The HEX code of the UKRI orange.
}
\description{
Access the HEX code of the orange used by the UKRI.
}
\examples{

barplot(
  1:3,
  col = ukri_orange()
)

}
\seealso{
Other UKRI Colours: 
\code{\link{ukri_dark_blue}()},
\code{\link{ukri_dark_purple}()},
\code{\link{ukri_light_blue}()},
\code{\link{ukri_light_purple}()},
\code{\link{ukri_mid_blue}()},
\code{\link{ukri_mid_purple}()}
}
\concept{UKRI Colours}
